R-NaD reg-value stream: scalar reg_v_head on the privileged value embedding, V_reg = propagated -eta*KL(pi||pi_reg) — reward = the EXPECTED penalty per state (rnad.py b5bf0657's eta_reg_entropy folded into the value recursion; own-side only as before), scalar clipped-IS v-trace targets (targets.compute_reg_returns, rho=c=min(1,isr), lambda 0.8, f32, terminal bootstraps 0), MSE at player_reg_value_coef 0.5 — the Q label and BOTH residual bases bootstrap on V_win + V_reg so every cell's Q prices FUTURE reference penalties, while the immediate per-cell penalty stays analytic in ref_penalised_q exactly as rnad.py puts eta_log_policy in the NeuRD q_vr (the win reward alone stays in the label: the taken cell's own penalty in the label would double-count the analytic term) — q_for_policy clip widens +-1 -> +-2 (Q now carries the reg stream; a row clipping flat would hand NeuRD zeros) — player_reg_ema_rate 1e-4 -> 5e-5 (= delta_m 20k as one continuous EMA; single reg param set kept over rnad.py's prev/prev_ crossfade pair by decision — no 4th net; if the pgaijs6l-style NeuRD gradient spikes (22% of steps at the clip past 56k, p99 1.6k, driver: the raw log-ratio outrunning the 1e-4 reference) persist, the pre-decided fallback is clipping the per-cell log-ratio, not more params) — panels player_reg_value_mean / player_reg_reward_mean / player_loss_v_reg (reward_mean should track -eta*player_ref_kl; value_mean ~ reward x remaining length)

+++ rl/model/heads.py
@@ -536,7 +536,9 @@ class RegressionValueLogitHead(nn.Module):
 
     @nn.compact
     def __call__(self, x: jax.Array):
-        x = MLP(**self.cfg.mlp.to_dict())(x)
+        # f32 out: scalar regression targets (reg value, builder ev /
+        # conditional entropy) are all f32-MSE consumers.
+        x = MLP(**self.cfg.mlp.to_dict())(x).astype(jnp.float32)
         if getattr(self.cfg, "output_activation", None) is not None:
             x = self.cfg.output_activation(x)
         return RegressionValueHeadOutput(logits=x.squeeze(-1))
